Add writeAsStream
Untested as of yet.  This is the more difficult algorithm of the two in
that we have to juggle the out queue to re-start styles when closing an
inner style, so that ANSI escapes will work correctly.

I should circle back and write a buffered iterator for this, as well as
the upcoming writeAsTree, so that write-and-flush workloads are doable.