Test writeToStream, tweak comparison in out queue
It turns out we want ties to sort shortest to longest.

I went with offset + length on instinct, and it turns out this is why.

Of course, it could be done the other way, and in fact there's probably
more use of final() than the bit of length-specific logic, but.  Still.