Clone the in queue before writing
The intention of Obelizmo is to allow reuse of marked strings for a few
printing methods, such as at the terminal and as HTML.

If we pop the queue without cloning it, then it's effectively destroyed
and we can't do that.  I might add a destructive option for the case of
one-and-done, but it doesn't seem important right now.