Add tests for bounded writeTree
That covers the bases, as far as surface area is concerned.

I have it in mind that it would be nice to have a context-object writer
which can customize based on various things, but it's not clear how the
interface would work, and I'd rather have nothing, than something which
was speculatively written and might not be useful for that reason.