Rewrite writeAsStream in a term-compatible way
This algorithm isn't perfect: for perfect we'd need total tracking that
exceeds my interest levels, with stacks for every kind of effect in the
repertoire.

This will do a good job with non-garish, realistic use of colours, that
is all I'm actually interested in.  The data format is the most capable
part of this system, many things can be built with it.

Once I've built more of those many things, then, I can think about more
fancy pantsy.