Pointer-agnostic Writers
Previously, various printers had a baked-in assumption that they would
get a pointer as the Writer type, which is not actually a safe premise
here.  So I've added some extra type reflection to account for passing
structs directly, as well as pointers to same.

The -Once writing methods for XtermLineWriter are buggy, in that they
don't uphold the invariant that marker memory does not belong to them
and therefore not be freed at any point.  This is a solvable problem,
in the meantime, a note to that effect is added.