Seek and drop
Add two methods to Xterm printer: one which seeks forward to an index,
and one which drops a line.  It's actually three, there's a convenience
wrapper for dropping a counted number of lines as well.

Fixes some minor bugs here and there, and rehabilitates the printOnce
functions, albeit without testing them.  They'll get redeployed in Vril,
so it didn't seem as important to add unit tests for them.

Obelizmo is now basically ready for debut, I want to take a pass to get
better organization into the color module, add some format printer for
user convenience, and add a bit more documentation.  But the module is
quite usable as is, and it's time to kick it out the door.